requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifultable==0.8.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-redis==5.2.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kavenegar==1.1.2
kombu==5.2.2
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.0.2
requests==2.26.0
requests-toolbelt==0.9.1
retrying==1.3.3
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
whitenoise==5.3.0
wrapt==1.13.3
